+++ src-tauri/src/models/ship/requirements.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Requirements {
-  pub power: Option<i64>,
-  pub crew: Option<i64>,
-  pub slots: Option<i64>
+  pub power: Option<i32>,
+  pub crew: Option<i32>,
+  pub slots: Option<i32>
 }
